qqoosie/requirements.txt
ffmpeg-python
librosa
pydub
pyfiglet
typer
yt-dlp
rich
soundfile
asciimatics
